+++ influxql/parser.go
@@ -687,8 +687,8 @@ func (p *Parser) parseSelectStatement(tr targetRequirement) (*SelectStatement, e
 		}
 	})
 
-	if d, _ := stmt.GroupByInterval(); stmt.IsRawQuery && d > 0 {
-		return nil, fmt.Errorf("GROUP BY requires at least one aggregate function")
+	if err := stmt.Validate(tr); err != nil {
+		return nil, err
 	}
 
 	return stmt, nil